refactor: 컴포넌트 명칭 변경(SelectBoxSection -> RestaurantListFilter)
