+++ index.html
@@ -9,7 +9,7 @@
 
   <body id="app">
     <nav-bar></nav-bar>
-    <select-box-section></select-box-section>
+    <restaurant-list-filter></restaurant-list-filter>
+    <add-restaurant-modal></add-restaurant-modal>
   </body>
-  <add-restaurant-modal></add-restaurant-modal>
 </html>
